move validation schema to page file
