+++ apps/pano/src/features/zod/utils.ts
@@ -1,0 +1,11 @@
+import type { z } from "zod";
+
+export type inferSafeParseErrors<
+  T extends z.ZodType<any, any, any>,
+  U = string
+> = {
+  formErrors: U[];
+  fieldErrors: {
+    [P in keyof z.infer<T>]?: U[];
+  };
+};
