+++ apps/pano/src/routes/settings.tsx
@@ -13,11 +13,8 @@ import type { ActionFunction, LoaderFunction } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import { useActionData } from "@remix-run/react";
 import type { FC } from "react";
-import type {
-  SettingsFields,
-  SettingsFieldsErrors,
-} from "~/features/zod/schemas";
-import { SchemaWithoutPasswords, SettingsSchema } from "~/features/zod/schemas";
+import { z } from "zod";
+import type { inferSafeParseErrors } from "~/features/zod/utils";
 import {
   updateEmail,
   updatePassword,
@@ -25,6 +22,23 @@ import {
 } from "~/models/user.server";
 import { requireUser } from "~/session.server";
 import { useUser } from "~/utils";
+
+const SettingsSchema = z.object({
+  oldPassword: z.string().min(6, { message: "Parolanızı eksik doldurdunuz" }),
+  newPassword: z
+    .string()
+    .min(6, { message: "Yeni parola en az 6 karakterden oluşmalıdır" }),
+  username: z.string().min(2, { message: "Kullanıcı adı uygun değil" }),
+  email: z.string().email({ message: "Geçerli bir email adresi girin" }),
+});
+
+const SchemaWithoutPasswords = SettingsSchema.omit({
+  oldPassword: true,
+  newPassword: true,
+});
+
+type SettingsFields = z.infer<typeof SettingsSchema>;
+type SettingsFieldsErrors = inferSafeParseErrors<typeof SettingsSchema>;
 
 export const loader: LoaderFunction = async ({ request }) => {
   await requireUser(request);
